% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxisymmetricTriangularElement.R
\name{AxisymmetricT_Area}
\alias{AxisymmetricT_Area}
\title{Area of an axisymmetric triangular element}
\usage{
AxisymmetricT_Area(vec_nodalcoordinates)
}
\arguments{
\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2,x3,y3).}
}
\value{
Area of an axisymmetric triangular triangular element.
}
\description{
This function computes the area of an axisymmetric triangular element
                                    from the coordinates of its nodes.
}
