% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxisymmetricTriangularElement.R
\name{AxisymmetricT_BF}
\alias{AxisymmetricT_BF}
\title{Equivalent body load for an element with body force}
\usage{
AxisymmetricT_BF(DOF = 6, Rb, Zb, vec_nodalcoordinates)
}
\arguments{
\item{DOF}{Degree of freedom of the element (6 by default)}

\item{Rb}{Radial body force.}

\item{Zb}{Axial body force}

\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2,x3,y3).}
}
\value{
A vector of equivalent nodal loads for an axisymmetric
                                                     triangular element with body load.
}
\description{
Equivalent body load for an element with body force
}
