% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxisymmetricTriangularElement.R
\name{AxisymmetricT_Centroid}
\alias{AxisymmetricT_Centroid}
\title{Centroid of an axisymmetric triangular element}
\usage{
AxisymmetricT_Centroid(vec_nodalcoordinates)
}
\arguments{
\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2,x3,y3).}
}
\value{
A vector (rbar and zbar) describing the centroid
                                         of an axisymmetric triangular element.
}
\description{
Centroid of an axisymmetric triangular element
}
