% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxisymmetricTriangularElement.R
\name{AxisymmetricT_Element_Matrix}
\alias{AxisymmetricT_Element_Matrix}
\title{Element stiffness matrix (axisymmetric triangular element)}
\usage{
AxisymmetricT_Element_Matrix(DOF = 6, YoungMod, Nu, vec_nodalcoordinates)
}
\arguments{
\item{DOF}{Degree of freedom (6 for an axisymmetric triangular element).}

\item{YoungMod}{Young's modulus.}

\item{Nu}{Poisson's ratio.}

\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2,x3,y3).}
}
\value{
Stiffness matrix of an axisymmetric triangular element.
}
\description{
Element stiffness matrix (axisymmetric triangular element)
}
