% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxisymmetricTriangularElement.R
\name{AxisymmetricT_ExpandedSFBF}
\alias{AxisymmetricT_ExpandedSFBF}
\title{Expanded vector of equivalent nodal loads (surface/body)}
\usage{
AxisymmetricT_ExpandedSFBF(TDOF, LoadColumnMatrix, i, j, k)
}
\arguments{
\item{TDOF}{Total degree of freedom.}

\item{LoadColumnMatrix}{The unexpanded vector of equivalent loads,
from either AxisymmetricT_BF() or AxisymmetricT_SF().}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
Expanded vector (a column matrix) of equivalent loads.
}
\description{
Expanded vector of equivalent nodal loads (surface/body)
}
