% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxisymmetricTriangularElement.R
\name{AxisymmetricT_Stresses}
\alias{AxisymmetricT_Stresses}
\title{Centroidal stress}
\usage{
AxisymmetricT_Stresses(YoungMod, Nu, vec_nodalcoord, vec_globalnodaldisp, i, j,
  k)
}
\arguments{
\item{YoungMod}{Young's modulus.}

\item{Nu}{Poisson's ratio.}

\item{vec_nodalcoord}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2,x3,y3).}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
Centroidal stresses in an axisymmetric triangular element.
}
\description{
Centroidal stress
}
