% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearBarElement.R
\name{Bar_Element_Forces}
\alias{Bar_Element_Forces}
\title{Local element forces (linear bar)}
\usage{
Bar_Element_Forces(DOF = 2, stiffbar, vec_globalnodaldisp, i, j)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{stiffbar}{Stiffness of the bar in the form of AE/L.}

\item{vec_globalnodaldisp}{Vector of global displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
Local nodal forces (linear bar).
}
\description{
This function generates the local forces at the nodes of a bar element.
}
