% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearBarElement.R
\name{Bar_Element_Matrix}
\alias{Bar_Element_Matrix}
\title{Element stiffness matrix (for a linear bar)}
\usage{
Bar_Element_Matrix(DOF = 2, stiffbar)
}
\arguments{
\item{DOF}{Degree of freedom for a bar element (2 for a linear bar).}

\item{stiffbar}{Stiffness of the bar in the form of AE/L.}
}
\value{
The 2 by 2 stiffness matrix.
}
\description{
This function generates the matrix for a linear bar element.
The axis of the bar element is assumed to coicide with the global x-axis.
}
