% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearBarElement.R
\name{Bar_ExpandedElement_Matrix}
\alias{Bar_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (for the linear bar)}
\usage{
Bar_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of bars.}

\item{eMatrix}{The 2 by 2 stiffness matrix of a specific bar element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix (conforming to the total degree of freedom of the system).
}
\description{
This function generates the expanded matrix for each element in a
connected system of bars.
}
