% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearBarElement.R
\name{Bar_Global_Forces}
\alias{Bar_Global_Forces}
\title{Global element forces (linear bar)}
\usage{
Bar_Global_Forces(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix.}

\item{vec_globalnodaldisp}{Vector of global nodal displacements.}
}
\value{
Global nodal forces (linear bar)
}
\description{
This function generates the global forces at the nodes of a bar element.
}
