% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearBarElement.R
\name{BarUDL_Matrix}
\alias{BarUDL_Matrix}
\title{Equivalent load vector for a UDL (bar element)}
\usage{
BarUDL_Matrix(DOF = 2, LoadMagnitude, Length)
}
\arguments{
\item{DOF}{Degree of freedom (2 for a bar).}

\item{LoadMagnitude}{Magnitude of the UDL, e.g. q in N/m.}

\item{Length}{Length}
}
\value{
A column matrix of the equivalent nodal loads.
}
\description{
Generates the load column matrix for an element with a
                            uniformly distributed load.
}
