% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerBeamElement.R
\name{BeamUDL_ExpandedMatirx}
\alias{BeamUDL_ExpandedMatirx}
\title{Expanded vector of the equivalent load}
\usage{
BeamUDL_ExpandedMatirx(TDOF, LoadColumnMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom.}

\item{LoadColumnMatrix}{The unexpanded vector of equivalent loads.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
Expanded vector (a column matrix) of equivalent loads.
}
\description{
This function generates the expanded vector of the equivalent load
                                            (Euler-Bernoulli beam).
}
