% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerBeamElement.R
\name{BeamUDL_Matrix}
\alias{BeamUDL_Matrix}
\title{Equivalent load vector for a UDL (beam element)}
\usage{
BeamUDL_Matrix(DOF = 4, LoadMagnitude, Length)
}
\arguments{
\item{DOF}{Degree of freedom (4 for a beam).}

\item{LoadMagnitude}{Magnitude of the UDL, e.g. q in N/m.}

\item{Length}{Length}
}
\value{
A column matrix of the equivalent nodal loads.
}
\description{
Generates the load column matrix for an element with a
                            uniformly distributed load.
}
