% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTriangularElement.R
\name{CSTriangular_Area}
\alias{CSTriangular_Area}
\title{Area of triangular element}
\usage{
CSTriangular_Area(vec_nodalcoordinates)
}
\arguments{
\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2,x3,y3).}
}
\value{
Area of a triangular element.
}
\description{
This function computes the area of a CST element
               from the coordinates of its nodes.
}
