% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTriangularElement.R
\name{CSTriangular_ExpandedElement_Matrix}
\alias{CSTriangular_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (linear triangular)}
\usage{
CSTriangular_ExpandedElement_Matrix(TDOF, eMatrix, i, j, k)
}
\arguments{
\item{TDOF}{Total degree of freedom of a discretized structure.}

\item{eMatrix}{The 6 by 6 stiffness matrix of a
specific linear triangular element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
The expanded matrix of a linear triangular element.
}
\description{
This function generates the expanded matrix for each element in a
connected system of linear triangular elements.
}
