% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTriangularElement.R
\name{CSTriangular_ExpandedSF}
\alias{CSTriangular_ExpandedSF}
\title{Expanded vector of equivalent load}
\usage{
CSTriangular_ExpandedSF(TDOF, LoadColumnMatrix, i, j, k)
}
\arguments{
\item{TDOF}{Total degree of freedom.}

\item{LoadColumnMatrix}{The unexpanded vector of equivalent loads.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
Expanded vector (a column matrix) of equivalent loads.
}
\description{
This function generates the expanded vector of equivalent nodal loads
                                            (linear triangular element).
}
