% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTriangularElement.R
\name{CSTriangular_GlobalForces}
\alias{CSTriangular_GlobalForces}
\title{Global nodal forces}
\usage{
CSTriangular_GlobalForces(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}
}
\value{
Global nodal forces.
}
\description{
This function generates the nodal global forces for linear triangular element.
}
