% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTriangularElement.R
\name{CSTriangular_LocalForces}
\alias{CSTriangular_LocalForces}
\title{Local element forces (linear triangular element)}
\usage{
CSTriangular_LocalForces(ematrix, vec_globalnodaldisp, i, j, k)
}
\arguments{
\item{ematrix}{Element matrix.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
Local nodal forces (linear triangular element).
}
\description{
Local element forces (linear triangular element)
}
