% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTriangularElement.R
\name{CSTriangular_SF}
\alias{CSTriangular_SF}
\title{Equivalent surface load for an element with distributed load}
\usage{
CSTriangular_SF(DOF = 6, SFtensile, SFshear, Length, thickness, case)
}
\arguments{
\item{DOF}{Degree of freedom of the element (6 by default)}

\item{SFtensile}{Magnitude of a uniform surface tensile pressure, e.g. q in N/m^2}

\item{SFshear}{Magnitude of a uniform surface shear pressure.}

\item{Length}{Lenth of the side on which load is acting.}

\item{thickness}{Thickness of the element}

\item{case}{Use 1 (if load is on side i-j,
2 if on side j-k, &
3 if on side i-k)}
}
\value{
A vector of equivalent loads for a CST element with distributed load.
}
\description{
Equivalent surface load for an element with distributed load
}
