% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearTriangularElement.R
\name{CSTriangular_Stresses}
\alias{CSTriangular_Stresses}
\title{Centroidal stress}
\usage{
CSTriangular_Stresses(YoungMod, Nu, thickness, vec_nodalcoord, case,
  vec_globalnodaldisp, i, j, k)
}
\arguments{
\item{YoungMod}{Young's modulus.}

\item{Nu}{Poisson's ratio.}

\item{thickness}{Thickness.}

\item{vec_nodalcoord}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2,x3,y3).}

\item{case}{Use 1 (for plane stress) or 2 for plane strain.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
Centroidal stresses in a linear triangular element.
}
\description{
Centroidal stress
}
