% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBeamElement.R
\name{Dynamic_NaturalFrequencies}
\alias{Dynamic_NaturalFrequencies}
\title{Natural frequencies (beam element)}
\usage{
Dynamic_NaturalFrequencies(reducedM, reducedK)
}
\arguments{
\item{reducedM}{Reduced mass matrix obtained by applying
boundary condition on the global mass matrix}

\item{reducedK}{Reduced mass matrix obtained by applying
boundary condition on the global stiffness matrix}
}
\value{
Natural frequencies in radian/s
}
\description{
his function computes the natural frequencies, in radians per seconds.
}
