% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBarElement.R
\name{DynamicBar_ExpandedElement_Matrix}
\alias{DynamicBar_ExpandedElement_Matrix}
\title{Expanded element matrix (dynamic bar)}
\usage{
DynamicBar_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of bars.}

\item{eMatrix}{The 2 by 2 stiffness or mass matrix of a
specific bar element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of a bar element.
}
\description{
This function returns the expanded
stiffness/mass matrix of a dynamic bar element.
}
