% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBarElement.R
\name{DynamicBar_StiffnessMatrix}
\alias{DynamicBar_StiffnessMatrix}
\title{Element stiffness matrix (Dynamic bar)}
\usage{
DynamicBar_StiffnessMatrix(DOF = 2, YoungMod, Area, Length)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{YoungMod}{Young's modulus.}

\item{Area}{Cross-sectional area}

\item{Length}{Length.}
}
\value{
Stiffness matrix of a dynamic bar element.
}
\description{
This function generates the 2 by 2 stiffness matrix of a bar element.
}
