% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBeamElement.R
\name{DynamicEulerBeam_ExpandedMatrix}
\alias{DynamicEulerBeam_ExpandedMatrix}
\title{Expanded element matrix (dynamic beam)}
\usage{
DynamicEulerBeam_ExpandedMatrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of beams.}

\item{eMatrix}{The 4 by 4 stiffness or mass matrix of a
specific beam element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of a bar element.
}
\description{
This function returns the expanded
stiffness/mass matrix of a dynamic beam element.
}
