% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBeamElement.R
\name{DynamicEulerBeam_MassMatrix}
\alias{DynamicEulerBeam_MassMatrix}
\title{Element mass matrix (Dynamic beam)}
\usage{
DynamicEulerBeam_MassMatrix(DOF = 4, Density, Area, Length)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{Density}{Material density}

\item{Area}{Cross-sectional area}

\item{Length}{Length.}
}
\value{
Mass matrix of a dynamic beam element.
}
\description{
This function generates the 4 by 4 consistent mass matrix of a beam element.
}
