% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBeamElement.R
\name{DynamicEulerBeam_StiffnessMatrix}
\alias{DynamicEulerBeam_StiffnessMatrix}
\title{Element stiffness matrix (Dynamic beam)}
\usage{
DynamicEulerBeam_StiffnessMatrix(DOF = 4, YoungMod, MomentI, Length)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{YoungMod}{Young's modulus.}

\item{MomentI}{Second moment of area.}

\item{Length}{Length.}
}
\value{
Stiffness matrix of a dynamic bar element.
}
\description{
This function generates the 4 by 4 stiffness matrix of a beam element.
}
