% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicPlaneTrussElement.R
\name{DynamicPlaneTruss_ExpandedMatrix}
\alias{DynamicPlaneTruss_ExpandedMatrix}
\title{Expanded stiffness/mass matrix (plane truss)}
\usage{
DynamicPlaneTruss_ExpandedMatrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom of a connected system of plane trusses.}

\item{eMatrix}{The 4 by 4 stiffness matrix of a
specific plane truss element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of a plane truss element.
}
\description{
This function generates the expanded matrix for each element in a
connected system of plane truss structure.
}
