% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicPlaneTrussElement.R
\name{DynamicPlaneTruss_LengthOrientation}
\alias{DynamicPlaneTruss_LengthOrientation}
\title{Geometric length and orientation from the xy coordinates (plane truss)}
\usage{
DynamicPlaneTruss_LengthOrientation(vec_nodalcoordinates)
}
\arguments{
\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2).}
}
\value{
Vector of the geometric length and orientation angle.
}
\description{
This function is used to get the geometric length  and orientation
from the nodal coordinates of an element.
}
