% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicPlaneTrussElement.R
\name{DynamicPlaneTruss_MassMatrix}
\alias{DynamicPlaneTruss_MassMatrix}
\title{Element mass matrix (dynamic plane truss)}
\usage{
DynamicPlaneTruss_MassMatrix(DOF = 4, Density, Area, Length, theta)
}
\arguments{
\item{DOF}{Degree of freedom (4 for a truss).}

\item{Density}{Material density}

\item{Area}{Cross-sectional area}

\item{Length}{Length}

\item{theta}{Angular orientation of a truss element.}
}
\value{
Mass matrix of a plane truss element.
}
\description{
This function generates the mass matrix for a plane truss element.
}
