% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicPlaneTrussElement.R
\name{DynamicPlaneTruss_StiffnessMatrix}
\alias{DynamicPlaneTruss_StiffnessMatrix}
\title{Element stiffness matrix (dynamic plane truss)}
\usage{
DynamicPlaneTruss_StiffnessMatrix(DOF = 4, stifftruss, theta)
}
\arguments{
\item{DOF}{Degree of freedom (4 for a truss).}

\item{stifftruss}{Axial stiffness of an element (N/m, i.e EA/L).}

\item{theta}{Angular orientation of a truss element.}
}
\value{
Stiffness matrix of a plane truss element.
}
\description{
This function generates the 4 by 4 stiffness matrix for
a plane truss element.
}
