% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerBeamElement.R
\name{EulerBeam_Element_Matrix}
\alias{EulerBeam_Element_Matrix}
\title{Element stiffness matrix (Euler-Bernoulli Beam)}
\usage{
EulerBeam_Element_Matrix(DOF = 4, YoungMod, MomentI, Length)
}
\arguments{
\item{DOF}{Degree of freedom (4 for a beam).}

\item{YoungMod}{Young modulus.}

\item{MomentI}{Principal moment of inertia of the cross-section.}

\item{Length}{Length}
}
\value{
Stiffness matrix of an Euler-Bernoulli beam element.
}
\description{
This function generates the 4 by 4 stiffness matrix for
an Euler-Bernoulli beam element.
}
