% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerBeamElement.R
\name{EulerBeam_ExpandedElement_Matrix}
\alias{EulerBeam_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (Euler-Bernoulli beam)}
\usage{
EulerBeam_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of beams.}

\item{eMatrix}{The 4 by 4 stiffness matrix of a
specific beam element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of an Euler-Bernoulli beam element.
}
\description{
This function generates the expanded matrix for each element in a
connected system of beams.
}
