% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerBeamElement.R
\name{EulerBeamMaxMin_BendingStress}
\alias{EulerBeamMaxMin_BendingStress}
\title{Bending stress (Euler-Bernoulli beam)}
\usage{
EulerBeamMaxMin_BendingStress(YoungMod, Length, vec_globalnodaldisp)
}
\arguments{
\item{YoungMod}{Young's modulus.}

\item{Length}{Length of an element.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}
}
\value{
A vector of bending stresses.
}
\description{
This function computes the maximum
   and minimum bending stress for a beam element.
}
