% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearSpringElement.R
\name{LinearSpring_Element_Forces}
\alias{LinearSpring_Element_Forces}
\title{Local element forces (spring)}
\usage{
LinearSpring_Element_Forces(DOF = 2, stiffk, vec_globalnodaldisp, i, j)
}
\arguments{
\item{DOF}{Degree of freedom for a spring element (2 by default)}

\item{stiffk}{Spring stiffness k in N/m}

\item{vec_globalnodaldisp}{Vector of global displacements}

\item{i}{Index of the first node}

\item{j}{Index of the second node}
}
\value{
Local nodal forces (spring)
}
\description{
Local element forces (spring)
}
