% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearSpringElement.R
\name{LinearSpring_Element_Matrix}
\alias{LinearSpring_Element_Matrix}
\title{Element stiffness matrix (for a spring)}
\usage{
LinearSpring_Element_Matrix(DOF = 2, stiffk)
}
\arguments{
\item{stiffk}{Spring stiffness k in N/m}
}
\value{
Stiffness matrix of a spring element
}
\description{
This function generates the matrix for a spring element.
The axis of the spring is assumed to coicide with the global x-axis.
}
