% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearSpringElement.R
\name{LinearSpring_ExpandedElement_Matrix}
\alias{LinearSpring_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (for a spring)}
\usage{
LinearSpring_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of springs}

\item{eMatrix}{Pre-expanded matrix of a specific spring element}

\item{i}{Index of the first node}

\item{j}{Index of the second node}
}
\value{
The expanded matrix (conforming to the total degree of freedom of the system)
}
\description{
This function generates the expanded matrix for each element in a connected system
of springs
}
