% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearSpringElement.R
\name{LinearSpring_Global_Forces}
\alias{LinearSpring_Global_Forces}
\title{Global element forces (spring)}
\usage{
LinearSpring_Global_Forces(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix}

\item{vec_globalnodaldisp}{Global nodal displacements}
}
\value{
Global nodal forces (spring)
}
\description{
Global element forces (spring)
}
