% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanarFrameElement.R
\name{PlaneFrame_Element_Matrix}
\alias{PlaneFrame_Element_Matrix}
\title{Element stiffness matrix (plane frame)}
\usage{
PlaneFrame_Element_Matrix(DOF = 6, YoungMod, Area, momentI, Length, theta)
}
\arguments{
\item{DOF}{Degree of freedom (6 for a plane frame).}

\item{YoungMod}{Young's modulus.}

\item{Area}{Cross-sectional area.}

\item{momentI}{Principal moment of inertia.}

\item{Length}{Element's geometric length.}

\item{theta}{Element's angular orientation.}
}
\value{
Stiffness matrix of a plane frame element.
}
\description{
This function generates the 6 by 6 stiffness matrix for
a plane frame element.
}
