% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanarFrameElement.R
\name{PlaneFrame_ExpandedElement_Matrix}
\alias{PlaneFrame_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (plane frame)}
\usage{
PlaneFrame_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom of a connected system of frame structures.}

\item{eMatrix}{The 6 by 6 stiffness matrix of a
specific plane frame element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of a plane frame element.
}
\description{
This function generates the expanded matrix for each element in a
connected system of plane frame structures.
}
