% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanarFrameElement.R
\name{PlaneFrame_GlobalForcesMoments}
\alias{PlaneFrame_GlobalForcesMoments}
\title{Global nodal forces and moments}
\usage{
PlaneFrame_GlobalForcesMoments(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}
}
\value{
Global nodal forces and moments.
}
\description{
This function generates the nodal global forces
                   and moments for plane frame element.
}
