% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanarFrameElement.R
\name{PlaneFrame_LocalForcesMoments}
\alias{PlaneFrame_LocalForcesMoments}
\title{Local element forces (plane frame)}
\usage{
PlaneFrame_LocalForcesMoments(YoungMod, Area, momentI, Length, theta,
  vec_globalnodaldisp, i, j)
}
\arguments{
\item{YoungMod}{Young's modulus.}

\item{Area}{Cross-sectional area.}

\item{momentI}{Principal moment of inertia.}

\item{Length}{Element's geometric length.}

\item{theta}{Element's orientation angle.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
Local nodal forces (plane frame).
}
\description{
Local element forces (plane frame)
}
