% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanarFrameElement.R
\name{PlaneFrameUDL_Matrix}
\alias{PlaneFrameUDL_Matrix}
\title{Equivalent load vector for a UDL (frame element)}
\usage{
PlaneFrameUDL_Matrix(DOF = 6, AxialUDLMagnitude, TransUDLMagnitude, Length,
  theta)
}
\arguments{
\item{DOF}{Degree of freedom (4 for a beam).}

\item{AxialUDLMagnitude}{Magnitude of the axial UDL, e.g. q in N/m.}

\item{TransUDLMagnitude}{Magnitude of the transverse UDL, e.g. q in N/m.}

\item{Length}{Element's geometric length.}

\item{theta}{Element's orientation angle.}
}
\value{
A column matrix of the equivalent nodal loads.
}
\description{
Generates the load column matrix for a plane frame element with a
                            uniformly distributed load.
}
