% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaneTrussElement.R
\name{PlaneTruss_AxialStress}
\alias{PlaneTruss_AxialStress}
\title{Axial Stress (plane truss)}
\usage{
PlaneTruss_AxialStress(YoungMod, elem_len, theta, vec_globalnodaldisp, i, j)
}
\arguments{
\item{YoungMod}{Young's modulus.}

\item{elem_len}{Length of an element.}

\item{theta}{Angular orientation of an element.}

\item{vec_globalnodaldisp}{Vector of global nodal displacement.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
Axial stress of a plane truss element.
}
\description{
This function generates the axial stress in a plane truss element.
}
