% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaneTrussElement.R
\name{PlaneTruss_Element_Matrix}
\alias{PlaneTruss_Element_Matrix}
\title{Element stiffness matrix (plane truss)}
\usage{
PlaneTruss_Element_Matrix(DOF = 4, stifftruss, theta)
}
\arguments{
\item{DOF}{Degree of freedom (4 for a truss).}

\item{stifftruss}{Axial stiffness of an element (N/m, i.e EA/L).}

\item{theta}{Angular orientation of a truss element.}
}
\value{
Stiffness matrix of a plane truss element.
}
\description{
This function generates the 4 by 4 stiffness matrix for
a plane truss element.
}
