% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaneTrussElement.R
\name{PlaneTruss_GeometricLengths}
\alias{PlaneTruss_GeometricLengths}
\title{Vector of geometric lengths}
\usage{
PlaneTruss_GeometricLengths(num_ele, vec_xnodalcoordinates,
  vec_ynodalcoordinates)
}
\arguments{
\item{num_ele}{Number of elements.}

\item{vec_xnodalcoordinates}{Vector of x coordinates.}

\item{vec_ynodalcoordinates}{Vector of y coordinates.}
}
\value{
Geometric lengths of elements from a set of coordinates.
}
\description{
This function returns a vector of geometric lengths for
a set of elements.
}
