% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaneTrussElement.R,
%   R/ThermalPlaneTrussElement.R
\name{PlaneTruss_LengthOrientation}
\alias{PlaneTruss_LengthOrientation}
\alias{PlaneTruss_LengthOrientation}
\title{Geometric length and orientation from xy coordinates (plane truss)}
\usage{
PlaneTruss_LengthOrientation(vec_nodalcoordinates)

PlaneTruss_LengthOrientation(vec_nodalcoordinates)
}
\arguments{
\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2).}

\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,x2,y2).}
}
\value{
Vector of the geometric length and orientation angle.

Vector of the geometric length and orientation angle.
}
\description{
This function is used to get the geometric length  and orientation
from the nodal coordinates of an element.
See also PlaneTruss_GeometricLengths() for lengths
                              of multiple elements.

This function generates the geometric length  and orientation
from the nodal coordinates of an element.
}
