% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaneTrussElement.R
\name{PlaneTruss_Orientations}
\alias{PlaneTruss_Orientations}
\title{Vector of angular orientations (plane truss)}
\usage{
PlaneTruss_Orientations(num_ele, vec_xnodalcoordinates, vec_ynodalcoordinates)
}
\arguments{
\item{num_ele}{Number of elements}

\item{vec_xnodalcoordinates}{Vector of x coordinates.}

\item{vec_ynodalcoordinates}{Vector of y coordinates.}
}
\value{
Angle of elements from a set of coordinates.
}
\description{
This function returns a vector of angular orientations for
a set of elements.
}
