% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticBarElement.R
\name{QuadraticBar_Element_Forces}
\alias{QuadraticBar_Element_Forces}
\title{Local element forces (quadratic bar)}
\usage{
QuadraticBar_Element_Forces(DOF, stiff, vec_globalnodaldisp, i, j, k)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{stiff}{Stiffness of the bar in the form of AE/L.}

\item{vec_globalnodaldisp}{Vector of global displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
Local nodal forces (quadratic bar).
}
\description{
This function generates the local forces at the nodes of
a quadratic bar element.
}
