% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticBarElement.R
\name{QuadraticBar_Element_Matrix}
\alias{QuadraticBar_Element_Matrix}
\title{Element stiffness matrix (quadratic bar)}
\usage{
QuadraticBar_Element_Matrix(DOF = 3, stiffbar)
}
\arguments{
\item{DOF}{Degree of freedom for a bar element (3 for a linear bar).}

\item{stiffbar}{Stiffness of the bar in the form of AE/L.}
}
\value{
The 3 by 3 stiffness matrix.
}
\description{
This function generates the matrix for a quadratic bar element.
The axis of the bar element is assumed to coicide with the global x-axis.
}
