% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticBarElement.R
\name{QuadraticBar_ExpandedElement_Matrix}
\alias{QuadraticBar_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (for the quadratic bar element)}
\usage{
QuadraticBar_ExpandedElement_Matrix(TDOF, eMatrix, i, j, k)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of bars.}

\item{eMatrix}{The 3 by 3 stiffness matrix of a specific quadratic bar element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}
}
\value{
The expanded matrix (conforming to the total degree of freedom of the system).
}
\description{
This function generates the expanded matrix for each element in a
connected system of bars.
}
