% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticBarElement.R
\name{QuadraticBar_Global_Forces}
\alias{QuadraticBar_Global_Forces}
\title{Global element forces (quadratic bar)}
\usage{
QuadraticBar_Global_Forces(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix.}

\item{vec_globalnodaldisp}{Vector of global nodal displacements.}
}
\value{
Global nodal forces (quadratic bar)
}
\description{
This function generates the global forces at the nodes of a quadratic bar element.
}
