% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearSpringElement.R
\name{ReducedLoadVector}
\alias{ReducedLoadVector}
\title{Reduced force vector (a column matrix of known force)}
\usage{
ReducedLoadVector(forcevector)
}
\arguments{
\item{forcevector}{Vector of known loads (be consistent with units)}
}
\value{
Reduced force vector
}
\description{
Reduced force vector (a column matrix of known force)
}
