% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearSpringElement.R
\name{ReducedStiffnessMatrix}
\alias{ReducedStiffnessMatrix}
\title{Reduced stiffness matrix}
\usage{
ReducedStiffnessMatrix(bigKmatrix, knownloadnodes)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix}

\item{knownloadnodes}{Vector of nodal index of known loads
in the form c(i,j,k)}
}
\value{
Reduced stiffness matrix
}
\description{
This function computes the reduced stiffness matrix
from the global stiffness matrix using the boundary condition of the system.
}
