% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceFrameElement.R
\name{SpaceFrame_ExpandedElement_Matrix}
\alias{SpaceFrame_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (space frame)}
\usage{
SpaceFrame_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom of a connected system
of space frame structures.}

\item{eMatrix}{The 12 by 12 stiffness matrix of a
specific space frame element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of a space frame element.
}
\description{
This function generates the expanded matrix for each element in a
connected system of space frame structures.
}
