% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTrussElement.R
\name{Spacetruss_AxialStress}
\alias{Spacetruss_AxialStress}
\title{Axial Stress (space truss)}
\usage{
Spacetruss_AxialStress(YoungMod, elem_len, thetas, vec_globalnodaldisp)
}
\arguments{
\item{YoungMod}{Young modulus.}

\item{elem_len}{Length of an element.}

\item{thetas}{Angular orientations of the element.}

\item{vec_globalnodaldisp}{Vector of global nodal displacement.}
}
\value{
Axial stress of a space truss element.
}
\description{
This function generates the axial stress in a space truss element.
}
