% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTrussElement.R
\name{SpaceTruss_Element_Matrix}
\alias{SpaceTruss_Element_Matrix}
\title{Element stiffness matrix (space truss)}
\usage{
SpaceTruss_Element_Matrix(DOF = 6, stifftruss, thetas)
}
\arguments{
\item{DOF}{Degree of freedom (6 for a space truss).}

\item{stifftruss}{Axial stiffness of an element (N/m, i.e EA/L).}

\item{thetas}{Angular orientation of a truss element.}
}
\value{
Stiffness matrix of a space truss element.
}
\description{
This function generates the 6 by 6 stiffness matrix for
a space truss element.
}
