% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTrussElement.R
\name{SpaceTruss_ExpandedElement_Matrix}
\alias{SpaceTruss_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (space truss)}
\usage{
SpaceTruss_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of truss.}

\item{eMatrix}{The 6 by 6 stiffness matrix of a specific truss element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of a space truss element.
}
\description{
This function generates the expanded matrix for each element in a
connected system of space truss structure.
}
