% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTrussElement.R
\name{SpaceTruss_LengthOrientations}
\alias{SpaceTruss_LengthOrientations}
\title{Geometric length and orientations from xyz coordinates (space truss)}
\usage{
SpaceTruss_LengthOrientations(vec_nodalcoordinates)
}
\arguments{
\item{vec_nodalcoordinates}{Vector of nodal coordinates
in the form c(x1,y1,z1,x2,y2,z3).}
}
\value{
Vector of the geometric length and orientation angleector of the geometric length and orientation angle.
}
\description{
This function is used to get the geometric length  and orientation
from the nodal coordinates of a space truss element.
}
