% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceTrussElement.R
\name{SpaceTruss_Local_Forces}
\alias{SpaceTruss_Local_Forces}
\title{Local element forces (space truss)}
\usage{
SpaceTruss_Local_Forces(YoungMod, Area, thetas, vec_globalnodaldisp, i, j)
}
\arguments{
\item{YoungMod}{Young modulus.}

\item{Area}{Cross-sectional area.}

\item{thetas}{Angular orientation of an element.}

\item{vec_globalnodaldisp}{Vector of global nodal displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the first node.}
}
\value{
Local nodal forces (space truss).
}
\description{
This function generates the local forces at the nodes of
 a space truss element.
}
