% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalBarElement.R
\name{ThermalBar_BF}
\alias{ThermalBar_BF}
\title{Thermal body force (thermal bar element)}
\usage{
ThermalBar_BF(DOF = 2, YoungMod, area, alpha, tempchange)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{YoungMod}{Young's modulus.}

\item{area}{Cross section area of a bar element.}

\item{alpha}{Thermal conductivity.}

\item{tempchange}{Temperature change.}
}
\value{
A vector of thermal body force for a bar element.
}
\description{
This function generates a vector of body force
for a bar element in a thermal analysis.
}
