% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalBarElement.R
\name{ThermalBar_Element_Forces}
\alias{ThermalBar_Element_Forces}
\title{Local nodal forces (thermal bar element)}
\usage{
ThermalBar_Element_Forces(DOF = 2, stiffbar, nodaldisp)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{stiffbar}{Stiffness of a bar element.}

\item{nodaldisp}{Element nodal displacement}
}
\value{
A vector of the nodal forces for
                         a bar element in a thermal analysis.
}
\description{
This function generates the nodal forces for a bar element
                                       in a thermal analysis.
}
