% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalBarElement.R
\name{ThermalBar_Element_Matrix}
\alias{ThermalBar_Element_Matrix}
\title{Element stiffness matrix (Bar element for thermal analysis)}
\usage{
ThermalBar_Element_Matrix(DOF = 2, stiffbar)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{stiffbar}{Young's modulus.}
}
\value{
Stiffness matrix of a thermal bar element.
}
\description{
This function generates the 2 by 2 stiffness matrix for
a thermal analysis involving bar elements.
}
