% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalBarElement.R
\name{ThermalBar_ExpandedBF}
\alias{ThermalBar_ExpandedBF}
\title{Expanded vector of thermal body force (thermal bar element)}
\usage{
ThermalBar_ExpandedBF(TDOF, LoadColumnMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of bars.}

\item{LoadColumnMatrix}{Vector of body force returned by ThermalBar_BF()}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded vector of a bar element.
}
\description{
This function generates the expanded vector of body force
for a bar element in a thermal analysis.
}
