% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalBarElement.R
\name{ThermalBar_ExpandedElement_Matrix}
\alias{ThermalBar_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (thermal bar element)}
\usage{
ThermalBar_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a
connected system of bars.}

\item{eMatrix}{The 2 by 2 stiffness matrix of a
specific bar element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded stiffness matrix of a
                                thermal bar element.
}
\description{
This function generates the expanded stiffness matrix
                                    for a bar element.
}
