% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalBarElement.R
\name{ThermalBar_Global_Forces}
\alias{ThermalBar_Global_Forces}
\title{Global nodal forces (thermal bar element)}
\usage{
ThermalBar_Global_Forces(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}
}
\value{
Vector of global nodal forces
                               (thermal bar element).
}
\description{
This function generates the global nodal forces for a bar element
                                       in a thermal analysis.
}
