% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalBarElement.R
\name{ThermalBar_Stresses}
\alias{ThermalBar_Stresses}
\title{Axial Stress (thermal bar element)}
\usage{
ThermalBar_Stresses(YoungMod, alpha, tempchange, Length, vec_globalnodaldisp, i,
  j)
}
\arguments{
\item{YoungMod}{Young's modulus.}

\item{alpha}{Thermal conductivity.}

\item{tempchange}{Temperature change.}

\item{Length}{Length of a bar element.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
A vector of thermal stress in a bar element.
}
\description{
This function generates the axial stress in a bar element
in a thermal analysis.
}
