% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalPlaneTrussElement.R
\name{ThermalPlaneTruss_BF}
\alias{ThermalPlaneTruss_BF}
\title{Thermal body force (thermal plane truss element)}
\usage{
ThermalPlaneTruss_BF(DOF = 4, YoungMod, area, alpha, tempchange, theta)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{YoungMod}{Young's modulus.}

\item{area}{Cross section area of a plane truss element.}

\item{alpha}{Thermal conductivity.}

\item{tempchange}{Temperature change.}

\item{theta}{Angular orientation of a truss element.}
}
\value{
A vector of thermal body force
                           for a plane truss element.
}
\description{
This function generates a vector of the body force
for a plane truss element in a thermal analysis.
}
