% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalPlaneTrussElement.R
\name{ThermalPlaneTruss_Element_Matrix}
\alias{ThermalPlaneTruss_Element_Matrix}
\title{Element stiffness matrix (thermal plane truss)}
\usage{
ThermalPlaneTruss_Element_Matrix(DOF = 4, stifftruss, theta)
}
\arguments{
\item{DOF}{Degree of freedom (4 for a truss).}

\item{stifftruss}{Axial stiffness of an element (N/m, i.e EA/L).}

\item{theta}{Angular orientation of a truss element.}
}
\value{
Stiffness matrix of a plane truss element
                        for thermal analysis.
}
\description{
This function generates the 4 by 4 stiffness matrix for
a plane truss elements in a thermal analysis.
}
