% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalPlaneTrussElement.R
\name{ThermalPlaneTruss_ExpandedBF}
\alias{ThermalPlaneTruss_ExpandedBF}
\title{Expanded vector of thermal body force (thermal plane truss element)}
\usage{
ThermalPlaneTruss_ExpandedBF(TDOF, LoadColumnMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected
system of plane truss element.}

\item{LoadColumnMatrix}{Vector of body force returned
by ThermalPlaneTruss_BF()}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded vector of a plane truss element.
}
\description{
This function generates the expanded vector of body force
for a plane truss element in a thermal analysis.
}
