% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalPlaneTrussElement.R
\name{ThermalPlaneTruss_ExpandedElement_Matrix}
\alias{ThermalPlaneTruss_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (thermal plane element)}
\usage{
ThermalPlaneTruss_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a
connected system of plane truss elements.}

\item{eMatrix}{The 4 by 4 stiffness matrix of a
specific bar element.}

\item{i}{Index of the first node.}

\item{j}{Index of the first node.}
}
\value{
The expanded stiffness matrix of a
                               thermal plane truss element.
}
\description{
This function generates the expanded stiffness matrix
        for plane truss element in a thermal analysis.
}
