% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalPlaneTrussElement.R
\name{ThermalPlaneTruss_Global_Forces}
\alias{ThermalPlaneTruss_Global_Forces}
\title{Global nodal forces (thermal plane truss)}
\usage{
ThermalPlaneTruss_Global_Forces(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}
}
\value{
Vector of global nodal forces
                               (thermal plane truss element).
}
\description{
This function generates the global nodal
forces for a plane truss element in a thermal analysis.
}
