% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermalPlaneTrussElement.R
\name{ThermalPlaneTruss_Local_Forces}
\alias{ThermalPlaneTruss_Local_Forces}
\title{Local nodal forces (thermal plane truss)}
\usage{
ThermalPlaneTruss_Local_Forces(stifftruss, theta, vec_globalnodaldisp)
}
\arguments{
\item{stifftruss}{Axial stiffness of the truss.}

\item{theta}{Angular orientation of an element.}

\item{vec_globalnodaldisp}{Vector of global nodal displacements.}
}
\value{
A vector of local nodal forces
                             for a plane truss element in a thermal analysis.
}
\description{
This function generates the local nodal forces for
a plane truss element in a thermal analysis.
}
