% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimoshenkoBeamElement.R
\name{Timoshenko_Element_Matrix}
\alias{Timoshenko_Element_Matrix}
\title{Element stiffness matrix (Timoshenko beam)}
\usage{
Timoshenko_Element_Matrix(DOF = 4, YoungModulus, MomentInertia, ShearArea,
  PoissonRatio, L = 1)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{YoungModulus}{Young's modulus.}

\item{MomentInertia}{Moment of inertia.}

\item{ShearArea}{Shear Area.}

\item{PoissonRatio}{Poisson's ratio.}

\item{L}{Length.}
}
\value{
Stiffness matrix of a Timoshenko beam element.
}
\description{
This function generates the 4 by 4 stiffness matrix for
a Timoshenko beam element.
}
