% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimoshenkoBeamElement.R
\name{Timoshenko_ExpandedElement_Matrix}
\alias{Timoshenko_ExpandedElement_Matrix}
\title{Expanded stiffness matrix (Timoshenko)}
\usage{
Timoshenko_ExpandedElement_Matrix(TDOF, eMatrix, i, j)
}
\arguments{
\item{TDOF}{Total degree of freedom in a connected system of beams.}

\item{eMatrix}{The 4 by 4 stiffness matrix of a
specific beam element.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
The expanded matrix of a Timoshenko beam element.
}
\description{
This function returns the expanded
stiffness matrix of a Timoshenko element.
}
