% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimoshenkoBeamElement.R
\name{Timoshenko_Global_ForcesMoments}
\alias{Timoshenko_Global_ForcesMoments}
\title{Global nodal forces and moments}
\usage{
Timoshenko_Global_ForcesMoments(bigKmatrix, vec_globalnodaldisp)
}
\arguments{
\item{bigKmatrix}{Global stiffness matrix.}

\item{vec_globalnodaldisp}{Vector of all global nodal displacements.}
}
\value{
Vector of global nodal forces and moments (Timoshenko beam element).
}
\description{
This function generates the nodal global forces
                   and moments for a Timoshenko beam element.
}
