% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimoshenkoBeamElement.R
\name{Timoshenko_Local_ForcesMoments}
\alias{Timoshenko_Local_ForcesMoments}
\title{Local forces and moments (Timoshenko)}
\usage{
Timoshenko_Local_ForcesMoments(DOF = 4, YoungModulus, MomentInertia,
  ShearArea, PoissonRatio, Length, vec_globalnodaldisp, vec_distributedLoad, i,
  j)
}
\arguments{
\item{DOF}{Degree of freedom.}

\item{YoungModulus}{Young's modulus.}

\item{MomentInertia}{Moment of inertia.}

\item{ShearArea}{Shear Area.}

\item{PoissonRatio}{Poisson's ratio.}

\item{Length}{Length.}

\item{vec_globalnodaldisp}{Vector of global nodal displacements.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{vec_distriload}{Vector of equivalent loads, if any.}
}
\value{
Local forces and moments (Timoshenko).
}
\description{
This function generates the nodal local forces
                   and moments for the element.
}
